+++ crates/cxx-qt/src/lib.rs
@@ -118,6 +118,11 @@ pub trait Threading: Sized {
     fn threading_drop(cxx_qt_thread: &mut CxxQtThread<Self>);
 }
 
+/// Placeholder for upcasting objects, suppresses dead code warning
+#[allow(dead_code)]
+#[doc(hidden)]
+pub trait Upcast<T> {}
+
 /// This trait can be implemented on any [CxxQtType] to define a
 /// custom constructor in C++ for the QObject.
 ///